requirements/base.txt
pydantic~=2.5.3
python-dotenv~=1.0.0
aiogram~=3.2.0
Babel~=2.14.0
babel~=2.13.1